+++ packages/office-ui-fabric-react/src/components/Facepile/Facepile.test.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import * as renderer from 'react-test-renderer';
 import { mount, shallow } from 'enzyme';
 import { setRTL } from '../../Utilities';
-import { Facepile } from './Facepile';
+import { FacepileBase } from './Facepile.base';
 import { IFacepilePersona, OverflowButtonType } from './Facepile.types';
 import { PersonaSize } from '../../Persona';
 import { Persona } from '../../Persona';
@@ -36,14 +36,14 @@ describe('Facepile', () => {
   });
 
   it('renders Facepile correctly', () => {
-    const component = renderer.create(<Facepile personas={ facepilePersonas } />);
+    const component = renderer.create(<FacepileBase personas={ facepilePersonas } />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders with only add button if no personas found and addButtonProps are not null', () => {
     const wrapper = mount(
-      <Facepile
+      <FacepileBase
         personas={ [] }
         addButtonProps={ {} }
         showAddButton={ true }
@@ -56,7 +56,7 @@ describe('Facepile', () => {
 
   it('renders chevron overflow button if overflowButtonProps are not null and OverflowButtonType equals downArrow', () => {
     const wrapper = mount(
-      <Facepile
+      <FacepileBase
         personas={ [] }
         overflowButtonProps={ {} }
         overflowButtonType={ OverflowButtonType.downArrow }
@@ -69,7 +69,7 @@ describe('Facepile', () => {
 
   it('renders more overflow button if overflowButtonProps are not null as OverflowButtonType equals more', () => {
     const wrapper = mount(
-      <Facepile
+      <FacepileBase
         personas={ [] }
         overflowButtonProps={ {} }
         overflowButtonType={ OverflowButtonType.more }
@@ -82,7 +82,7 @@ describe('Facepile', () => {
 
   it('renders without descriptive overflow button if overflowButtonProps are not null and maximum personas are not exceeded', () => {
     const wrapper = mount(
-      <Facepile
+      <FacepileBase
         personas={ [] }
         overflowButtonProps={ {} }
         overflowButtonType={ OverflowButtonType.descriptive }
@@ -96,7 +96,7 @@ describe('Facepile', () => {
   it('renders with descriptive overflow button if overflowButtonProps are not null and maximum personas are exceeded', () => {
     const personas: IFacepilePersona[] = facepilePersonas.concat(...facepilePersonas, ...facepilePersonas);
     const wrapper = mount(
-      <Facepile
+      <FacepileBase
         personas={ personas }
         maxDisplayablePersonas={ 5 }
         overflowButtonProps={ {} }
@@ -109,7 +109,7 @@ describe('Facepile', () => {
 
   it('renders no more than maximum allowed personas', () => {
     const wrapper = mount(
-      <Facepile
+      <FacepileBase
         personas={ facepilePersonas.concat(facepilePersonas, facepilePersonas, facepilePersonas) }
         maxDisplayablePersonas={ 2 }
       />
@@ -128,7 +128,7 @@ describe('Facepile', () => {
     }];
 
     const wrapper = mount(
-      <Facepile
+      <FacepileBase
         personas={ personas }
       />
     );
@@ -142,7 +142,7 @@ describe('Facepile', () => {
 
   it('personas and buttons render default size if not specified', () => {
     const wrapper = mount(
-      <Facepile
+      <FacepileBase
         personas={ facepilePersonas }
         addButtonProps={ {} }
         showAddButton={ true }
@@ -166,7 +166,7 @@ describe('Facepile', () => {
   it('personas and buttons render specified size', () => {
     // Test XXS size renders
     let wrapper = shallow(
-      <Facepile
+      <FacepileBase
         personas={ facepilePersonas }
         personaSize={ PersonaSize.size24 }
       />);
@@ -179,7 +179,7 @@ describe('Facepile', () => {
 
     // Test small size renders
     wrapper = shallow(
-      <Facepile
+      <FacepileBase
         personas={ facepilePersonas }
         personaSize={ PersonaSize.size40 }
       />);
@@ -193,7 +193,7 @@ describe('Facepile', () => {
 
   it('renders Persona control if exactly one persona is sent in props', () => {
     let wrapper = shallow(
-      <Facepile
+      <FacepileBase
         personas={ facepilePersonas.slice(0, 1) }
         overflowPersonas={ [] }
       />);
@@ -202,7 +202,7 @@ describe('Facepile', () => {
     expect(wrapper.find(Persona).length).toEqual(1);
 
     wrapper = shallow(
-      <Facepile
+      <FacepileBase
         personas={ facepilePersonas.slice(0, 1) }
       />);
 
@@ -212,7 +212,7 @@ describe('Facepile', () => {
 
   it('renders Persona control if exactly one persona is sent in props', () => {
     let wrapper = shallow(
-      <Facepile
+      <FacepileBase
         personas={ facepilePersonas.slice(0, 1) }
         overflowPersonas={ [] }
       />);
@@ -221,7 +221,7 @@ describe('Facepile', () => {
     expect(wrapper.find(Persona).length).toEqual(1);
 
     wrapper = shallow(
-      <Facepile
+      <FacepileBase
         personas={ facepilePersonas.slice(0, 1) }
       />);
