+++ packages/office-ui-fabric-react/src/components/Facepile/index.ts
@@ -1,2 +1,3 @@
 export * from './Facepile.types';
+export * from './Facepile.base';
 export * from './Facepile';
